# backend/shortVideoApiService/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /collection/video:
        get:
            tags:
                - CollectionService
            description: 列出收藏夹中的视频
            operationId: CollectionService_ListVideo4Collection
            parameters:
                - name: collectionId
                  in: query
                  description: '@gotags: json:"collectionId,omitempty,string"'
                  schema:
                    type: string
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.ListVideo4CollectionResponse'
        post:
            tags:
                - CollectionService
            description: 将视频添加到收藏夹
            operationId: CollectionService_AddVideo2Collection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.AddVideo2CollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.AddVideo2CollectionResponse'
        delete:
            tags:
                - CollectionService
            description: 从收藏夹中移除视频
            operationId: CollectionService_RemoveVideoFromCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.RemoveVideoFromCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.RemoveVideoFromCollectionResponse'
...
    /comment/video:
        post:
            tags:
                - CommentService
            description: 列出视频的评论
            operationId: CommentService_ListComment4Video
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.ListComment4VideoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.ListComment4VideoResponse'
    /cover/upload:
        post:
            tags:
                - ShortVideoCoreVideoService
            description: 预注册上传封面
            operationId: ShortVideoCoreVideoService_PreSign4UploadCover
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.PreSign4UploadRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.PreSign4UploadResponse'
...
    /file/{fileId}/finish:
        post:
            tags:
                - ShortVideoCoreVideoService
            description: 通用确认上传完成
            operationId: ShortVideoCoreVideoService_ReportFinishUpload
            parameters:
                - name: fileId
                  in: path
                  description: |-
                    文件id
                     @gotags: json:"file_id,omitempty,string"
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.ReportFinishUploadRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.ReportFinishUploadResponse'
...
    /video/feed:
        post:
            tags:
                - ShortVideoCoreVideoService
            description: 刷视频
            operationId: ShortVideoCoreVideoService_FeedShortVideo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.FeedShortVideoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.FeedShortVideoResponse'
    /video/finish:
        post:
            tags:
                - ShortVideoCoreVideoService
            description: 确认视频上传完成
            operationId: ShortVideoCoreVideoService_ReportVideoFinishUpload
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.ReportVideoFinishUploadRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.ReportVideoFinishUploadResponse'
    /video/list:
        post:
            tags:
                - ShortVideoCoreVideoService
            description: 获取当前用户的发布视频列表
            operationId: ShortVideoCoreVideoService_ListPublishedVideo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.ListPublishedVideoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.ListPublishedVideoResponse'
    /video/upload:
        post:
            tags:
                - ShortVideoCoreVideoService
            description: 预注册上传视频
            operationId: ShortVideoCoreVideoService_PreSign4UploadVideo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.PreSign4UploadVideoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.PreSign4UploadVideoResponse'
    /video/{videoId}:
        get:
            tags:
                - ShortVideoCoreVideoService
            description: 获取视频信息
            operationId: ShortVideoCoreVideoService_GetVideoById
            parameters:
                - name: videoId
                  in: path
                  description: '@gotags: json:"video_id,omitempty,string"'
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.GetVideoByIdResponse'
...
components:
    schemas:
...
        svapi.AddVideo2CollectionRequest:
            type: object
            properties:
                collectionId:
                    type: string
                    description: '@gotags: json:"collectionId,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
        svapi.AddVideo2CollectionResponse:
            type: object
            properties: {}
...
        svapi.Comment:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                parentId:
                    type: string
                    description: '@gotags: json:"parentId,omitempty,string"'
                user:
                    $ref: '#/components/schemas/svapi.CommentUser'
                replyUser:
                    $ref: '#/components/schemas/svapi.CommentUser'
                content:
                    type: string
                date:
                    type: string
                likeCount:
                    type: string
                replyCount:
                    type: string
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Comment'
        svapi.CommentUser:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...
        svapi.FeedShortVideoRequest:
            type: object
            properties:
                latestTime:
                    type: string
                    description: '@gotags: json:"latest_time,omitempty,string"'
                userId:
                    type: string
                    description: '@gotags: json:"user_id,omitempty,string"'
                feedNum:
                    type: string
                    description: '@gotags: json:"feed_num,omitempty,string"'
            description: 刷视频请求消息类型
        svapi.FeedShortVideoResponse:
            type: object
            properties:
                meta:
                    $ref: '#/components/schemas/svapi.Metadata'
                videos:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Video'
                nextTime:
                    type: string
                    description: '@gotags: json:"next_time,omitempty,string"'
            description: 刷视频响应消息类型
...
        svapi.GetVideoByIdResponse:
            type: object
            properties:
                video:
                    $ref: '#/components/schemas/svapi.Video'
            description: 获取视频信息响应消息类型
...
        svapi.ListComment4VideoRequest:
            type: object
            properties:
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationRequest'
        svapi.ListComment4VideoResponse:
            type: object
            properties:
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Comment'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationResponse'
...
        svapi.ListPublishedVideoRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"user_id,omitempty,string"'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationRequest'
            description: 获取当前用户的发布视频列表请求消息类型
        svapi.ListPublishedVideoResponse:
            type: object
            properties:
                videoList:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Video'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationResponse'
            description: 获取当前用户的发布视频列表响应消息类型
        svapi.ListVideo4CollectionResponse:
            type: object
            properties:
                videos:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Video'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationResponse'
...
        svapi.Metadata:
            type: object
            properties:
                bizCode:
                    type: integer
                    format: int32
                message:
                    type: string
                domain:
                    type: string
                reason:
                    type: array
                    items:
                        type: string
        svapi.PaginationRequest:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                size:
                    type: integer
                    format: int32
                sort:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.SortField'
        svapi.PaginationResponse:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                count:
                    type: integer
                    format: int32
        svapi.PreSign4UploadRequest:
            type: object
            properties:
                hash:
                    type: string
                    description: md5 hash
                fileType:
                    type: string
                    description: 文件类型，默认mp4
                size:
                    type: string
                    description: |-
                        文件大小，单位byte
                         @gotags: json:"size,omitempty,string"
                filename:
                    type: string
                    description: 文件名
            description: 通用小文件单次上传请求消息类型
        svapi.PreSign4UploadResponse:
            type: object
            properties:
                url:
                    type: string
                    description: minio上传地址
                fileId:
                    type: string
                    description: |-
                        文件id
                         @gotags: json:"file_id,omitempty,string"
            description: 通用小文件单次上传响应消息类型
        svapi.PreSign4UploadVideoRequest:
            type: object
            properties:
                hash:
                    type: string
                    description: md5 hash
                fileType:
                    type: string
                    description: 文件类型，默认mp4
                size:
                    type: string
                    description: |-
                        文件大小，单位byte
                         @gotags: json:"size,omitempty,string"
                filename:
                    type: string
                    description: 文件名
            description: 预注册上传视频请求消息类型
        svapi.PreSign4UploadVideoResponse:
            type: object
            properties:
                url:
                    type: string
                    description: minio上传地址
                fileId:
                    type: string
                    description: |-
                        文件id
                         @gotags: json:"file_id,omitempty,string"
            description: 预注册上传视频响应消息类型
...
        svapi.RemoveVideoFromCollectionRequest:
            type: object
            properties:
                collectionId:
                    type: string
                    description: '@gotags: json:"collectionId,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
        svapi.RemoveVideoFromCollectionResponse:
            type: object
            properties: {}
        svapi.ReportFinishUploadRequest:
            type: object
            properties:
                fileId:
                    type: string
                    description: |-
                        文件id
                         @gotags: json:"file_id,omitempty,string"
            description: 通用确认上传完成请求消息类型
        svapi.ReportFinishUploadResponse:
            type: object
            properties:
                url:
                    type: string
                    description: url
            description: 通用确认上传完成响应消息类型
...
        svapi.ReportVideoFinishUploadRequest:
            type: object
            properties:
                fileId:
                    type: string
                    description: |-
                        文件id
                         @gotags: json:"file_id,omitempty,string"
                title:
                    type: string
                    description: 视频标题
                coverUrl:
                    type: string
                    description: 视频封面地址
                description:
                    type: string
                    description: 视频描述
                videoUrl:
                    type: string
                    description: 视频地址
            description: 确认视频上传完成请求消息类型
        svapi.ReportVideoFinishUploadResponse:
            type: object
            properties:
                videoId:
                    type: string
                    description: |-
                        视频id
                         @gotags: json:"video_id,omitempty,string"
            description: 确认视频上传完成响应消息类型
        svapi.SortField:
            type: object
            properties:
                field:
                    type: string
                order:
                    type: integer
                    format: enum
...
        svapi.Video:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                author:
                    $ref: '#/components/schemas/svapi.VideoAuthor'
                playUrl:
                    type: string
                coverUrl:
                    type: string
                favoriteCount:
                    type: string
                    description: '@gotags: json:"favoriteCount,omitempty,string"'
                commentCount:
                    type: string
                    description: '@gotags: json:"commentCount,omitempty,string"'
                isFavorite:
                    type: boolean
                title:
                    type: string
                isCollected:
                    type: boolean
                collectedCount:
                    type: string
                    description: '@gotags: json:"collectedCount,omitempty,string"'
        svapi.VideoAuthor:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...